requirements.txt
celery
factory_boy
flask
flask-cors
Flask-Migrate
Flask-SQLAlchemy
Flask-Testing
pre-commit
psycopg2
pytest
python-dotenv
redis
requests
sqlalchemy-json
sqlalchemy-utils
